Update NumberController to support floating point and negative numbers; return 200 OK for valid inputs including invalid strings

+++ src/main/java/com/devops/numberclassification/api/dto/response/NumberResponse.java
@@ -22,7 +22,7 @@ import java.util.List;
 
 )
 public class NumberResponse {
-    private int number;
+    private double number; // Change from int to double
 
     @JsonProperty("is_prime")
     private boolean isPrime;
